+++ .eslintrc.js
@@ -21,7 +21,17 @@ module.exports = {
     "plugin:react/recommended",
     "plugin:prettier/recommended",
   ],
-  plugins: ["react", "import", "jest", "prettier"],
+  plugins: ["react", "import", "jest", "prettier", "@typescript-eslint"],
+  overrides: [
+    {
+      files: ["**/*.ts", "**/*.tsx"],
+      parser: "@typescript-eslint/parser",
+      parserOptions: {
+        ecmaFeatures: { jsx: true },
+      },
+      extends: ["plugin:@typescript-eslint/recommended"],
+    },
+  ],
   settings: {
     react: {
       pragma: "React",
