+++ package.json
@@ -76,7 +76,7 @@
   "dependencies": {
     "@babel/runtime-corejs3": "^7.27.1",
     "@scarf/scarf": "=1.4.0",
-    "@tanstack/react-virtual": "=3.14.9",
+    "@tanstack/react-virtual": "=3.14.10",
     "base64-js": "^1.5.1",
     "buffer": "^6.0.3",
     "classnames": "^2.5.1",
@@ -129,6 +129,8 @@
     "@pmmmwh/react-refresh-webpack-plugin": "^0.6.2",
     "@release-it/conventional-changelog": "=12.0.0",
     "@svgr/webpack": "=8.1.0",
+    "@typescript-eslint/eslint-plugin": "=8.67.0",
+    "@typescript-eslint/parser": "=8.67.0",
     "autoprefixer": "^10.5.4",
     "babel-loader": "^9.2.1",
     "babel-plugin-lodash": "=3.3.4",
